# swagger_server/models/applicant_req.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class ApplicantReq(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, industry: str=None, area: str=None):  # noqa: E501
        """ApplicantReq - a model defined in Swagger

        :param industry: The industry of this ApplicantReq.  # noqa: E501
        :type industry: str
        :param area: The area of this ApplicantReq.  # noqa: E501
        :type area: str
        """
        self.swagger_types = {
            'industry': str,
            'area': str
        }

        self.attribute_map = {
            'industry': 'industry',
            'area': 'area'
        }
        self._industry = industry
        self._area = area

    @classmethod
    def from_dict(cls, dikt) -> 'ApplicantReq':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ApplicantReq of this ApplicantReq.  # noqa: E501
        :rtype: ApplicantReq
        """
        return util.deserialize_model(dikt, cls)

    @property
    def industry(self) -> str:
        """Gets the industry of this ApplicantReq.


        :return: The industry of this ApplicantReq.
        :rtype: str
        """
        return self._industry

    @industry.setter
    def industry(self, industry: str):
        """Sets the industry of this ApplicantReq.


        :param industry: The industry of this ApplicantReq.
        :type industry: str
        """

        self._industry = industry

    @property
    def area(self) -> str:
        """Gets the area of this ApplicantReq.


        :return: The area of this ApplicantReq.
        :rtype: str
        """
        return self._area

    @area.setter
    def area(self, area: str):
        """Sets the area of this ApplicantReq.


        :param area: The area of this ApplicantReq.
        :type area: str
        """

        self._area = area