# swagger_server/models/applicant_res.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.applicant_res_data import ApplicantResData  # noqa: F401,E501
from swagger_server import util


class ApplicantRes(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, data: ApplicantResData=None):  # noqa: E501
        """ApplicantRes - a model defined in Swagger

        :param data: The data of this ApplicantRes.  # noqa: E501
        :type data: ApplicantResData
        """
        self.swagger_types = {
            'data': ApplicantResData
        }

        self.attribute_map = {
            'data': 'data'
        }
        self._data = data

    @classmethod
    def from_dict(cls, dikt) -> 'ApplicantRes':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ApplicantRes of this ApplicantRes.  # noqa: E501
        :rtype: ApplicantRes
        """
        return util.deserialize_model(dikt, cls)

    @property
    def data(self) -> ApplicantResData:
        """Gets the data of this ApplicantRes.


        :return: The data of this ApplicantRes.
        :rtype: ApplicantResData
        """
        return self._data

    @data.setter
    def data(self, data: ApplicantResData):
        """Sets the data of this ApplicantRes.


        :param data: The data of this ApplicantRes.
        :type data: ApplicantResData
        """

        self._data = data