# swagger_server/models/area_res.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.area_res_data import AreaResData  # noqa: F401,E501
from swagger_server import util


class AreaRes(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, data: AreaResData=None):  # noqa: E501
        """AreaRes - a model defined in Swagger

        :param data: The data of this AreaRes.  # noqa: E501
        :type data: AreaResData
        """
        self.swagger_types = {
            'data': AreaResData
        }

        self.attribute_map = {
            'data': 'data'
        }
        self._data = data

    @classmethod
    def from_dict(cls, dikt) -> 'AreaRes':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The AreaRes of this AreaRes.  # noqa: E501
        :rtype: AreaRes
        """
        return util.deserialize_model(dikt, cls)

    @property
    def data(self) -> AreaResData:
        """Gets the data of this AreaRes.


        :return: The data of this AreaRes.
        :rtype: AreaResData
        """
        return self._data

    @data.setter
    def data(self, data: AreaResData):
        """Sets the data of this AreaRes.


        :param data: The data of this AreaRes.
        :type data: AreaResData
        """

        self._data = data