# swagger_server/models/trend1_res.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.trend1_res_data import Trend1ResData  # noqa: F401,E501
from swagger_server import util


class Trend1Res(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, data: Trend1ResData=None):  # noqa: E501
        """Trend1Res - a model defined in Swagger

        :param data: The data of this Trend1Res.  # noqa: E501
        :type data: Trend1ResData
        """
        self.swagger_types = {
            'data': Trend1ResData
        }

        self.attribute_map = {
            'data': 'data'
        }
        self._data = data

    @classmethod
    def from_dict(cls, dikt) -> 'Trend1Res':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Trend1Res of this Trend1Res.  # noqa: E501
        :rtype: Trend1Res
        """
        return util.deserialize_model(dikt, cls)

    @property
    def data(self) -> Trend1ResData:
        """Gets the data of this Trend1Res.


        :return: The data of this Trend1Res.
        :rtype: Trend1ResData
        """
        return self._data

    @data.setter
    def data(self, data: Trend1ResData):
        """Sets the data of this Trend1Res.


        :param data: The data of this Trend1Res.
        :type data: Trend1ResData
        """

        self._data = data