# swagger_server/models/trend2_res.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.trend2_res_data import Trend2ResData  # noqa: F401,E501
from swagger_server import util


class Trend2Res(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, data: Trend2ResData=None):  # noqa: E501
        """Trend2Res - a model defined in Swagger

        :param data: The data of this Trend2Res.  # noqa: E501
        :type data: Trend2ResData
        """
        self.swagger_types = {
            'data': Trend2ResData
        }

        self.attribute_map = {
            'data': 'data'
        }
        self._data = data

    @classmethod
    def from_dict(cls, dikt) -> 'Trend2Res':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Trend2Res of this Trend2Res.  # noqa: E501
        :rtype: Trend2Res
        """
        return util.deserialize_model(dikt, cls)

    @property
    def data(self) -> Trend2ResData:
        """Gets the data of this Trend2Res.


        :return: The data of this Trend2Res.
        :rtype: Trend2ResData
        """
        return self._data

    @data.setter
    def data(self, data: Trend2ResData):
        """Sets the data of this Trend2Res.


        :param data: The data of this Trend2Res.
        :type data: Trend2ResData
        """

        self._data = data