# swagger_server/models/type_req.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class TypeReq(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, industry: str=None):  # noqa: E501
        """TypeReq - a model defined in Swagger

        :param industry: The industry of this TypeReq.  # noqa: E501
        :type industry: str
        """
        self.swagger_types = {
            'industry': str
        }

        self.attribute_map = {
            'industry': 'industry'
        }
        self._industry = industry

    @classmethod
    def from_dict(cls, dikt) -> 'TypeReq':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The TypeReq of this TypeReq.  # noqa: E501
        :rtype: TypeReq
        """
        return util.deserialize_model(dikt, cls)

    @property
    def industry(self) -> str:
        """Gets the industry of this TypeReq.


        :return: The industry of this TypeReq.
        :rtype: str
        """
        return self._industry

    @industry.setter
    def industry(self, industry: str):
        """Sets the industry of this TypeReq.


        :param industry: The industry of this TypeReq.
        :type industry: str
        """

        self._industry = industry