# swagger_server/models/type_res.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.type_res_data import TypeResData  # noqa: F401,E501
from swagger_server import util


class TypeRes(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, data: TypeResData=None):  # noqa: E501
        """TypeRes - a model defined in Swagger

        :param data: The data of this TypeRes.  # noqa: E501
        :type data: TypeResData
        """
        self.swagger_types = {
            'data': TypeResData
        }

        self.attribute_map = {
            'data': 'data'
        }
        self._data = data

    @classmethod
    def from_dict(cls, dikt) -> 'TypeRes':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The TypeRes of this TypeRes.  # noqa: E501
        :rtype: TypeRes
        """
        return util.deserialize_model(dikt, cls)

    @property
    def data(self) -> TypeResData:
        """Gets the data of this TypeRes.


        :return: The data of this TypeRes.
        :rtype: TypeResData
        """
        return self._data

    @data.setter
    def data(self, data: TypeResData):
        """Sets the data of this TypeRes.


        :param data: The data of this TypeRes.
        :type data: TypeResData
        """

        self._data = data